重要更新：重构 osu_user_info_archive，节约存储空间

+++ src/main/java/com/now/nowbot/model/osu/OsuUser.kt
@@ -506,10 +506,10 @@ open class OsuUser {
     val countryRank: Long
         get() = statistics?.countryRank ?: 0L
 
-    val levelCurrent: Int
+    val levelCurrent: Short
         get() = statistics?.levelCurrent ?: 0
 
-    val levelProgress: Int
+    val levelProgress: Short
         get() = statistics?.levelProgress ?: 0
 
     override fun toString(): String {
